Add support for registering/matching cmd wildcards
When the user registers a command, they can use the '*' atom at the end
of the command spec to allow matching any single string per each '*'
used.

This isn't very useful yet, since the details of exactly what the user
entered will not be passed back to the command callback. But fear not,
that functionality will come in a future commit!