Provide the user-inputted cmd to its callback
This allows commands with wildcard fields to have access to the actual
user-inputted strings. This follows on from the previous checkin,
allowing for us to implement callbacks for commands like "riak-admin
cluster join <node>" without having to use a cumbersome "--node=node"
flag to specify the node.

This also breaks the existing callback API, but it's the only clean way
I could find to do it. It's not really all that bad though: existing
command callbacks just need to add "_," to their function header to
ignore the extra argument we've added. Since we're still pre-1.0, this
seems like a reasonable compromise.