Fix clash with callbacks returning {error, E}
Just matching on a 2-tuple wasn't specific enough - callbacks are
allowed to return {error, _} which is supposed to cause a corresponding
message to be formatted and printed by clique.

We could make it more specific by adding an is_integer guard, but we
probably should have been explicitly matching on something specific like
this in the first place.