Modify "set" to use new config callback style
We no longer pass the flags through to the config change callback, and
the callback is executed directly on whichever node(s) are receiving the
config change. Additionally, the return format of the config callbacks
has changed so that we can display either a text status message for a
standard "set", or a table of results when we use the --all flag.